soc_planning_agent/requirements.txt
anthropic>=0.40.0
schedule>=1.2.2
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0
requests>=2.31.0
